requirements.txt
fpdf==1.7.2
numpy==1.24.4
opencv-python==4.9.0.80
pdf2image==1.17.0
pillow==10.3.0
pypdf==4.2.0